Fix config issues (credit to @jaacksondev)

+++ src/main/java/com/teamabnormals/allurement/core/mixin/ExperienceOrbMixin.java
@@ -22,7 +22,7 @@ public class ExperienceOrbMixin {
 	private void onCollideWithPlayer(Player player, CallbackInfo ci) {
 		int count = AllurementUtil.getTotalEnchantmentLevel(AllurementEnchantments.ALLEVIATING.get(), player, EquipmentSlot.Type.ARMOR);
 		if (count > 0) {
-			float factor = AllurementConfig.COMMON.alleviatingHealingFactor.get() * count;
+			float factor = AllurementConfig.COMMON.alleviatingHealingFactor.get().floatValue() * count;
 			float i = Math.min(this.value * factor, player.getMaxHealth() - player.getHealth());
 			this.value -= Math.round(i / factor);
 			player.heal(i);
